Я ВЕРЮ

+++ kaseev.andrey/S1/AdditionalFunctions.cpp
@@ -87,7 +87,7 @@ namespace kaseev {
   calculateSumList(const kaseev::List<std::pair<std::string, kaseev::List<int>>> &arr)
   {
     bool finished = false;
-    size_t index = 0;
+    int index = 0;
     kaseev::List<std::pair<std::string, kaseev::List<int>>> sum;
 
     while (!finished)
@@ -96,7 +96,7 @@ namespace kaseev {
       std::string currentString;
       int currentSum = 0;
 
-      for (size_t i = 0; i < arr.size(); ++i)
+      for (int i = 0; i < arr.size(); ++i)
       {
         const kaseev::List<int> &sublist = arr[i].second;
